[fix] 컴포넌트 이름 수정, HotCarousel Content 하나의 컴포넌트로 합침
