+++ dutchiepay/src/app/_components/_landing/Product_Hot.jsx
@@ -1,127 +1,218 @@
+// Product_Hot.jsx
 import '@/styles/commerce.css';
 import '@/styles/globals.css';
-
-import { useRef, useState, useEffect } from 'react';
+import Image from 'next/image';
+import Link from 'next/link';
+import clock from '../../../../public/image/clock.svg';
+import product from '../../../../public/image/product1.jpg';
 import Slider from 'react-slick';
-import TopFive_Content from './TopFive_Content';
-import BottomFive_Content from './BottomFive_Content';
-export default function Product_Hot() {
-  const handleHotSlideChange = (index) => {
-    setActiveSlide(index);
-    sliderWrapper.current.slickGoTo(index); // 올바른 슬라이드로 이동
-  };
-  const sliderWrapper = useRef(null);
-  const sliderTopFive = useRef(null);
-  const sliderBottomFive = useRef(null);
-  const [activeSlide, setActiveSlide] = useState(0);
-  const [activeSlideBottom, setActiveSlideBottom] = useState(0);
-  const [activeSlideTop, setActiveSlideTop] = useState(0);
-  const settingsTopFive = {
-    dots: false,
-    infinite: false,
-    speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    autoplay: true,
-    autoplaySpeed: 3000,
-    vertical: true,
-    beforeChange: (current, next) => setActiveSlideTop(next),
-    afterChange: (current) => {
-      if (current === 4) {
-        setTimeout(() => {
-          if (
-            sliderWrapper.current &&
-            sliderBottomFive.current &&
-            activeSlide === 0
-          ) {
-            sliderWrapper.current.slickGoTo(1);
-            sliderBottomFive.current.slickGoTo(0);
-            sliderBottomFive.current.slickPlay();
-          }
-        }, 3000);
-      }
-    },
-  };
-  useEffect(() => {
-    if (sliderTopFive.current) {
-      sliderTopFive.current.slickPlay();
-    }
-  }, []);
-  const settingsWrapper = {
-    dots: false,
-    infinite: false,
-    speed: 1000,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    autoplay: false,
-    vertical: true,
-    beforeChange: (current, next) => setActiveSlide(next),
-  };
 
-  const settingsBottomFive = {
-    dots: false,
-    infinite: false,
-    speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    autoplay: true,
-    autoplaySpeed: 3000,
-    vertical: true,
-    beforeChange: (current, next) => setActiveSlideBottom(next),
-    afterChange: (current) => {
-      if (current === 4) {
-        setTimeout(() => {
-          if (sliderWrapper.current && activeSlide === 1) {
-            sliderWrapper.current.slickGoTo(0); // 1~5위 슬라이더로 이동
-            sliderTopFive.current.slickGoTo(0);
-            sliderTopFive.current.slickPlay();
-          }
-        }, 1000);
-      }
-    },
-  };
-  //추후 데이터 들어오면 Link title 글자수 제한 코드 추가 필요
+export default function Product_Hot({
+  sliderTopFive,
+  settingsTopFive,
+  activeSlideTop,
+  sliderBottomFive,
+  activeSlideBottom,
+  settingsBottomFive,
+  sliderWrapper,
+  settingsWrapper,
+}) {
   return (
     <>
-      <div className="relative">
-        <h2 className="main__title">가장 HOT🔥한</h2>
-        <div className="absolute bottom-[0px] left-[900px] w-[200px]">
-          <button
-            onClick={() => handleHotSlideChange(0)}
-            className={`px-[8px] mx-[8px] ${
-              activeSlide === 0
-                ? 'border bg-blue--200 text-white rounded-full'
-                : ''
-            }`}
-          >
-            1
-          </button>
-          <button
-            onClick={() => handleHotSlideChange(1)}
-            className={`px-[8px] ${
-              activeSlide === 1
-                ? 'border bg-blue--200 text-white rounded-full'
-                : ''
-            }`}
-          >
-            2
-          </button>
+      <Slider ref={sliderWrapper} {...settingsWrapper}>
+        <div>
+          <div className="w-[1000px] h-[300px] px-[20px] flex flex-row justify-between items-center">
+            <div className="w-[470px]">
+              <Slider
+                ref={sliderTopFive}
+                {...settingsTopFive}
+                className="text-center"
+              >
+                {Array.from({ length: 5 }, (_, index) => (
+                  <div className="flex justify-center items-center" key={index}>
+                    <Link
+                      href="/commerce/detail?productId=123"
+                      title="애슐리 볶음밥 10인분 혼합 구성 10종"
+                      className="w-[480px] flex items-center"
+                    >
+                      <div className="w-[200px] h-[200px] relative">
+                        <Image
+                          className="object-cover"
+                          src={product}
+                          alt="애슐리 볶음밥"
+                          fill
+                        />
+                      </div>
+                      <div className="w-[240px] pl-[16px]">
+                        <div className="flex items-center gap-[4px]">
+                          <Image
+                            className="w-[18px] h-[18px]"
+                            src={clock}
+                            alt="남은 시간"
+                            width={18}
+                            height={18}
+                          />
+                          <p className="text-blue--700 text-sm font-semibold">
+                            12일 08시간 36분 남음
+                          </p>
+                        </div>
+                        <p className="title--multi-line text-lg font-semibold mt-[4px] text-left">
+                          {index + 1} 홈플래닛 반반 멀티 전기그릴 (개별 온도
+                          조절)
+                        </p>
+                        <div className="flex gap-[8px] items-center">
+                          <p className="text-[12px] text-gray--500 line-through">
+                            32,500원
+                          </p>
+                          <strong className="text-lg text-blue--500">
+                            27,500원
+                          </strong>
+                          <p className="bg-red--500 rounded-2xl text-white text-xs px-[4px] py-[2px]">
+                            30%
+                          </p>
+                        </div>
+                        <span className="flex justify-end items-baseline text-xs text-gray--500">
+                          공구 성공까지{' '}
+                          <strong className="text-3xl text-blue--500">
+                            &nbsp;15%
+                          </strong>
+                        </span>
+                      </div>
+                    </Link>
+                  </div>
+                ))}
+              </Slider>
+            </div>
+            <div>
+              <div className="w-[488px]">
+                <ul className="w-[488px] h-[200px] flex flex-col items-center">
+                  {Array.from({ length: 5 }, (_, index) => (
+                    <li
+                      className={`w-[488px] h-[40px] flex flex-row items-center gap-2 px-[20px] ${
+                        activeSlideTop === index
+                          ? 'border border-[#000] rounded-md'
+                          : ''
+                      }`}
+                      key={index}
+                    >
+                      <p
+                        className={`text-lg ${activeSlideTop === index ? 'text-[#333]' : 'text-[#999]'}`}
+                      >
+                        {index + 1}
+                      </p>
+                      <p
+                        className={`text-sm w-[100%] overflow-hidden text-ellipsis text-nowrap ${
+                          activeSlideTop === index
+                            ? 'text-[#333]'
+                            : 'text-[#999]'
+                        }`}
+                      >
+                        [{index + 1}] [한정기획] 에센허브 티트리 100 오일
+                      </p>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            </div>
+          </div>
         </div>
-      </div>
-      <div className="w-[1020px]  flex flex-row">
-        <Slider ref={sliderWrapper} {...settingsWrapper}>
-          <TopFive_Content
-            sliderTopFive={sliderTopFive}
-            settingsTopFive={settingsTopFive}
-            activeSlideTop={activeSlideTop}
-          />
-          <BottomFive_Content
-            sliderBottomFive={sliderBottomFive}
-            settingsBottomFive={settingsBottomFive}
-            activeSlideBottom={activeSlideBottom}
-          />
-        </Slider>
-      </div>
+        <div>
+          <div className="w-[1000px] h-[300px] px-[20px] flex flex-row justify-between items-center">
+            <div className="w-[470px]">
+              <Slider
+                ref={sliderBottomFive}
+                {...settingsBottomFive}
+                className="text-center"
+              >
+                {Array.from({ length: 5 }, (_, index) => (
+                  <div className="flex justify-center items-center" key={index}>
+                    <Link
+                      href="/commerce/detail?productId=123"
+                      title="애슐리 볶음밥 10인분 혼합 구성 10종"
+                      className="w-[480px] flex items-center"
+                    >
+                      <div className="w-[200px] h-[200px] relative">
+                        <Image
+                          className="object-cover"
+                          src={product}
+                          alt="애슐리 볶음밥"
+                          fill
+                        />
+                      </div>
+                      <div className="w-[240px] pl-[16px]">
+                        <div className="flex items-center gap-[4px]">
+                          <Image
+                            className="w-[18px] h-[18px]"
+                            src={clock}
+                            alt="남은 시간"
+                            width={18}
+                            height={18}
+                          />
+                          <p className="text-blue--700 text-sm font-semibold">
+                            12일 08시간 36분 남음
+                          </p>
+                        </div>
+                        <p className="title--multi-line text-lg font-semibold mt-[4px] text-left">
+                          {index + 6} 홈플래닛 반반 멀티 전기그릴 (개별 온도
+                          조절)
+                        </p>
+                        <div className="flex gap-[8px] items-center">
+                          <p className="text-[12px] text-gray--500 line-through">
+                            32,500원
+                          </p>
+                          <strong className="text-lg text-blue--500">
+                            27,500원
+                          </strong>
+                          <p className="bg-red--500 rounded-2xl text-white text-xs px-[4px] py-[2px]">
+                            30%
+                          </p>
+                        </div>
+                        <span className="flex justify-end items-baseline text-xs text-gray--500">
+                          공구 성공까지{' '}
+                          <strong className="text-3xl text-blue--500">
+                            &nbsp;15%
+                          </strong>
+                        </span>
+                      </div>
+                    </Link>
+                  </div>
+                ))}
+              </Slider>
+            </div>
+            <div>
+              <div className="w-[488px]">
+                <ul className="w-[488px] h-[200px] flex flex-col items-center">
+                  {Array.from({ length: 5 }, (_, index) => (
+                    <li
+                      className={`w-[488px] h-[40px] flex flex-row items-center gap-2 px-[20px] ${
+                        activeSlideBottom === index
+                          ? 'border border-[#000] rounded-md'
+                          : ''
+                      }`}
+                      key={index}
+                    >
+                      <p
+                        className={`text-lg ${activeSlideBottom === index ? 'text-[#333]' : 'text-[#999]'}`}
+                      >
+                        {index + 6}
+                      </p>
+                      <p
+                        className={`text-sm w-[100%] overflow-hidden text-ellipsis text-nowrap ${
+                          activeSlideBottom === index
+                            ? 'text-[#333]'
+                            : 'text-[#999]'
+                        }`}
+                      >
+                        [{index + 6}] [한정기획] 에센허브 티트리 100 오일
+                      </p>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            </div>
+          </div>
+        </div>
+      </Slider>
     </>
   );
 }
